first pass working

+++ src/commands/do_init.js
@@ -1,18 +1,21 @@
 import FS from 'async-file';
 import Path from 'path';
+import { CONFIG_FILENAME } from '../lib/project';
 
-const CONFIG_FILENAME = "migrations.config.js";
-
-export default async function doInit({ args }) {
-  const path = Path.join(process.cwd(), CONFIG_FILENAME);
-  await FS.writeFile(path, `export default {
-  migrationsTableName: "migrations",
-  migrationsPath: "./migrations/",
-  getConnection: async ()=>{
-    const pg = /* do stuff to get a single pg client to your database */
-    return pg;
-  }
+export default async function doInit(project) {
+  const path = Path.join(project.rootPath, CONFIG_FILENAME);
+  if (!(await FS.exists(path))) {
+    await FS.writeFile(path, `export default {
+migrationsTableName: "migrations",
+migrationsPath: "./migrations/",
+getConnection: async ()=>{
+  const pg = /* do stuff to get a single pg client to your database */
+  return pg;
 }
-`);
-  console.log(`Wrote skeleton to ${path}`);
+}
+  `);
+    console.log(`Wrote skeleton to ${path}`);
+  } else {
+    console.log(`Already configured at ${path}`);
+  }
 }
